Attempted to make dropdowns for dashboard
So many errors. I don't know how to fix.
here is a tutorial i followed.
tried for hours to make it work. It still does not work. It's telling me that my types are any and then it's telling me that my types are never and then it's telling me that it expects '...' and I tried and failed to fix.

https://www.youtube.com/watch?v=t8JK5bVoVBw&ab_channel=LogRocket